+++ example/example_auth.dart
@@ -30,14 +30,15 @@ void main() async {
   exit(0);
 }
 
-Future<SpotifyApi> _getUserAuthenticatedSpotifyApi(
+Future<SpotifyApi?> _getUserAuthenticatedSpotifyApi(
     SpotifyApiCredentials credentials) async {
   print(
       'Please paste your redirect url (from your spotify application\'s redirect url):');
   var redirect = stdin.readLineSync();
 
   var grant = SpotifyApi.authorizationCodeGrant(credentials);
-  var authUri = grant.getAuthorizationUrl(Uri.parse(redirect), scopes: _scopes);
+  var authUri =
+      grant.getAuthorizationUrl(Uri.parse(redirect!), scopes: _scopes);
 
   print(
       'Please paste this url \n\n$authUri\n\nto your browser and enter the redirected url:');
@@ -53,8 +54,8 @@ Future<SpotifyApi> _getUserAuthenticatedSpotifyApi(
   return SpotifyApi.fromClient(client);
 }
 
-void _currentlyPlaying(SpotifyApi spotify) async =>
-    await spotify.me.currentlyPlaying().then((a) {
+Future<void> _currentlyPlaying(SpotifyApi spotify) async =>
+    await spotify.me.currentlyPlaying().then((Player? a) {
       if (a == null || a.item == null) {
         print('Nothing currently playing.');
         return;
@@ -62,8 +63,8 @@ void _currentlyPlaying(SpotifyApi spotify) async =>
       print('Currently playing: ${a.item?.name}');
     }).catchError(_prettyPrintError);
 
-void _devices(SpotifyApi spotify) async =>
-    await spotify.me.devices().then((devices) {
+Future<void> _devices(SpotifyApi spotify) async =>
+    await spotify.me.devices().then((Iterable<Device>? devices) {
       if (devices == null || devices.isEmpty) {
         print('No devices currently playing.');
         return;
@@ -72,11 +73,10 @@ void _devices(SpotifyApi spotify) async =>
       print(devices.map((device) => device.name).join(', '));
     }).catchError(_prettyPrintError);
 
-void _followingArtists(SpotifyApi spotify) async {
+Future<void> _followingArtists(SpotifyApi spotify) async {
   var cursorPage = spotify.me.following(FollowingType.artist);
-  await cursorPage.all().then((cursorPage) {
-    var followingArtists = cursorPage.map((artist) => artist.name).join(', ');
-    print('Following artists: $followingArtists');
+  await cursorPage.first().then((cursorPage) {
+    print(cursorPage.items!.map((artist) => artist.name).join(', '));
   }).catchError((ex) => _prettyPrintError(ex));
 }
 
@@ -91,7 +91,7 @@ void _createPrivatePlaylist(SpotifyApi spotify) async {
   var user = await spotify.me.get();
   await spotify.playlists
       .createPlaylist(
-    user.id,
+    user.id!,
     'Cool New Playlist 2',
     description: 'Songs to test by',
     public: false,