+++ lib/src/models/paging.dart
@@ -6,13 +6,14 @@ part of spotify.models;
 typedef ParserFunction<T> = T Function(dynamic object);
 
 Iterable<dynamic> itemsNativeFromJson(List<dynamic> json) => json;
-List<Map> itemsNativeToJson(Iterable<dynamic> items) => List.from(items);
+List<Map> itemsNativeToJson(Iterable<dynamic>? items) =>
+    (items == null) ? [] : List.from(items);
 
 class BasePaging<T> extends Object {
   BasePaging();
 
   /// A link to the Web API endpoint returning the full result of the request.
-  String href;
+  String? href;
 
   /// The requested data
   ///
@@ -20,17 +21,17 @@ class BasePaging<T> extends Object {
   /// requested data as a deserialized list.
   @JsonKey(
       name: 'items', fromJson: itemsNativeFromJson, toJson: itemsNativeToJson)
-  Iterable<dynamic> itemsNative;
+  Iterable<dynamic>? itemsNative;
 
   /// The maximum number of items in the response (as set in the query or by
   /// default).
-  int limit;
+  int limit = 20;
 
   /// URL to the next page of items. ([null] if none)
-  String next;
+  String? next;
 
   /// The total number of items available to return.
-  int total;
+  int total = 0;
 }
 
 @JsonSerializable(createToJson: false)
@@ -40,10 +41,10 @@ class Paging<T> extends BasePaging<T> {
   factory Paging.fromJson(Map<String, dynamic> json) => _$PagingFromJson(json);
 
   /// The offset of the items returned (as set in the query or by default).
-  int offset;
+  int? offset;
 
   /// URL to the previous page of items. (null if none)
-  String previous;
+  String? previous;
 }
 
 @JsonSerializable(createToJson: false)