+++ lib/src/spotify.dart
@@ -4,9 +4,10 @@
 part of spotify;
 
 class SpotifyApi extends SpotifyApiBase {
-  SpotifyApi(SpotifyApiCredentials credentials, {Function(SpotifyApiCredentials) onCredentialsRefreshed})
+  SpotifyApi(SpotifyApiCredentials credentials,
+      {Function(SpotifyApiCredentials)? onCredentialsRefreshed})
       : super(credentials, http.Client(), onCredentialsRefreshed);
-  
+
   SpotifyApi.fromClient(FutureOr<oauth2.Client> client)
       : super.fromClient(client);
 
@@ -14,10 +15,24 @@ class SpotifyApi extends SpotifyApiBase {
       oauth2.AuthorizationCodeGrant grant, String responseUri)
       : super.fromAuthCodeGrant(grant, responseUri);
 
-  static oauth2.AuthorizationCodeGrant authorizationCodeGrant(
-      SpotifyApiCredentials credentials, {Function(SpotifyApiCredentials) onCredentialsRefreshed}) {
-    return SpotifyApiBase.authorizationCodeGrant(credentials, http.Client(), onCredentialsRefreshed);
+  SpotifyApi.withAccessToken(String accessToken)
+      : super._withAccessToken(accessToken);
+
+  static Future<SpotifyApi> asyncFromCredentials(
+    SpotifyApiCredentials credentials, {
+    Function(SpotifyApiCredentials)? onCredentialsRefreshed,
+  }) {
+    return SpotifyApiBase._asyncFromCredentials(
+      credentials,
+      http.Client(),
+      onCredentialsRefreshed,
+    );
   }
 
-   
+  static oauth2.AuthorizationCodeGrant authorizationCodeGrant(
+      SpotifyApiCredentials credentials,
+      {Function(SpotifyApiCredentials)? onCredentialsRefreshed}) {
+    return SpotifyApiBase.authorizationCodeGrant(
+        credentials, http.Client(), onCredentialsRefreshed);
+  }
 }