+++ lib/src/spotify_base.dart
@@ -6,36 +6,37 @@ part of spotify;
 abstract class SpotifyApiBase {
   static const String _baseUrl = 'https://api.spotify.com';
   static const String _tokenUrl = 'https://accounts.spotify.com/api/token';
-  static const String _authorizationUrl = 'https://accounts.spotify.com/authorize';
+  static const String _authorizationUrl =
+      'https://accounts.spotify.com/authorize';
 
   bool _shouldWait = false;
-  FutureOr<oauth2.Client> _client;
-  Artists _artists;
+  late FutureOr<oauth2.Client> _client;
+  late Artists _artists;
   Artists get artists => _artists;
-  Albums _albums;
+  late Albums _albums;
   Albums get albums => _albums;
-  Tracks _tracks;
+  late Tracks _tracks;
   Tracks get tracks => _tracks;
-  Playlists _playlists;
+  late Playlists _playlists;
   Playlists get playlists => _playlists;
-  RecommendationsEndpoint _recommendations;
+  late RecommendationsEndpoint _recommendations;
   RecommendationsEndpoint get recommendations => _recommendations;
-  Users _users;
+  late Users _users;
   Users get users => _users;
-  Search _search;
+  late Search _search;
   Search get search => _search;
-  AudioFeatures _audioFeatures;
+  late AudioFeatures _audioFeatures;
   AudioFeatures get audioFeatures => _audioFeatures;
-  Categories _categories;
+  late Categories _categories;
   Categories get categories => _categories;
-  Me _me;
+  late Me _me;
   Me get me => _me;
-  Shows _shows;
+  late Shows _shows;
   Shows get shows => _shows;
-  oauth2.Client get client => _client;
+  FutureOr<oauth2.Client> get client => _client;
 
   SpotifyApiBase.fromClient(FutureOr<http.BaseClient> client) {
-    _client = client;
+    _client = client as FutureOr<oauth2.Client>;
 
     _artists = Artists(this);
     _albums = Albums(this);
@@ -50,30 +51,72 @@ abstract class SpotifyApiBase {
     _shows = Shows(this);
   }
 
-  SpotifyApiBase(SpotifyApiCredentials credentials, [http.BaseClient httpClient, Function(SpotifyApiCredentials) callBack])
+  SpotifyApiBase(SpotifyApiCredentials credentials,
+      [http.Client? httpClient, Function(SpotifyApiCredentials)? callBack])
       : this.fromClient(_getOauth2Client(credentials, httpClient, callBack));
 
-  SpotifyApiBase.fromAuthCodeGrant(oauth2.AuthorizationCodeGrant grant, String responseUri)
-      : this.fromClient(grant.handleAuthorizationResponse(Uri.parse(responseUri).queryParameters));
+  SpotifyApiBase.fromAuthCodeGrant(
+      oauth2.AuthorizationCodeGrant grant, String responseUri)
+      : this.fromClient(grant.handleAuthorizationResponse(
+            Uri.parse(responseUri).queryParameters));
 
-  static oauth2.AuthorizationCodeGrant authorizationCodeGrant(SpotifyApiCredentials credentials, http.BaseClient httpClient,
-      [Function(SpotifyApiCredentials) callBack]) {
-    if (callBack == null)
-      return oauth2.AuthorizationCodeGrant(credentials.clientId, Uri.parse(SpotifyApiBase._authorizationUrl), Uri.parse(SpotifyApiBase._tokenUrl),
-          secret: credentials.clientSecret, httpClient: httpClient);
- 
-    return oauth2.AuthorizationCodeGrant(credentials.clientId, Uri.parse(SpotifyApiBase._authorizationUrl), Uri.parse(SpotifyApiBase._tokenUrl),
-          secret: credentials.clientSecret, httpClient: httpClient, onCredentialsRefreshed: (oauth2.Credentials cred) {
-        SpotifyApiCredentials newCredentials = SpotifyApiCredentials(credentials.clientId, credentials.clientSecret,
-            accessToken: cred.accessToken, expiration: cred.expiration, refreshToken: cred.refreshToken, scopes: cred.scopes);
-        callBack(newCredentials);
-      });
+  SpotifyApiBase._withAccessToken(String accessToken)
+      : this.fromClient(oauth2.Client(oauth2.Credentials(accessToken)));
+
+  static Future<SpotifyApi> _asyncFromCredentials(
+    SpotifyApiCredentials credentials, [
+    http.Client? httpClient,
+    Function(SpotifyApiCredentials)? callBack,
+  ]) async {
+    final client = await _getOauth2Client(
+      credentials,
+      httpClient,
+      callBack,
+    );
+
+    return SpotifyApi.fromClient(client);
   }
 
-  static FutureOr<oauth2.Client> _getOauth2Client(SpotifyApiCredentials credentials, http.BaseClient httpClient,
-      [Function(SpotifyApiCredentials) callBack]) async {
+  static oauth2.AuthorizationCodeGrant authorizationCodeGrant(
+      SpotifyApiCredentials credentials, http.Client httpClient,
+      [Function(SpotifyApiCredentials)? callBack]) {
+    return oauth2.AuthorizationCodeGrant(
+        credentials.clientId!,
+        Uri.parse(SpotifyApiBase._authorizationUrl),
+        Uri.parse(SpotifyApiBase._tokenUrl),
+        secret: credentials.clientSecret,
+        httpClient: httpClient,
+        onCredentialsRefreshed: callBack != null
+            ? (oauth2.Credentials cred) {
+                final newCredentials = SpotifyApiCredentials(
+                    credentials.clientId, credentials.clientSecret,
+                    accessToken: cred.accessToken,
+                    expiration: cred.expiration,
+                    refreshToken: cred.refreshToken,
+                    scopes: cred.scopes);
+                callBack(newCredentials);
+              }
+            : null);
+  }
+
+  static FutureOr<oauth2.Client> _getOauth2Client(
+      SpotifyApiCredentials credentials, http.Client? httpClient,
+      [Function(SpotifyApiCredentials)? callBack]) async {
     if (credentials.fullyQualified) {
       var oauthCredentials = credentials._toOauth2Credentials();
+
+      void credentialRefreshedWrapperCallback(oauth2.Credentials cred) {
+        callBack?.call(
+          SpotifyApiCredentials(
+            credentials.clientId,
+            credentials.clientSecret,
+            accessToken: cred.accessToken,
+            expiration: cred.expiration,
+            refreshToken: cred.refreshToken,
+            scopes: cred.scopes,
+          ),
+        );
+      }
 
       if (oauthCredentials.isExpired) {
         oauthCredentials = await oauthCredentials.refresh(
@@ -81,21 +124,15 @@ abstract class SpotifyApiBase {
           secret: credentials.clientSecret,
           httpClient: httpClient,
         );
+        credentialRefreshedWrapperCallback(oauthCredentials);
       }
 
       return oauth2.Client(
         oauthCredentials,
         identifier: credentials.clientId,
-        onCredentialsRefreshed: callBack == null
-            ? null
-            : (oauth2.Credentials cred) {
-                SpotifyApiCredentials newCredentials = SpotifyApiCredentials(credentials.clientId, credentials.clientSecret,
-                    accessToken: cred.accessToken, expiration: cred.expiration, refreshToken: cred.refreshToken, scopes: cred.scopes);
-                callBack(newCredentials);
-              },
+        onCredentialsRefreshed: credentialRefreshedWrapperCallback,
         secret: credentials.clientSecret,
       );
- 
     }
 
     return oauth2.clientCredentialsGrant(
@@ -107,30 +144,30 @@ abstract class SpotifyApiBase {
   }
 
   Future<String> _get(String path) {
-    return _getImpl('${_baseUrl}/$path', const {});
+    return _getImpl('$_baseUrl/$path', const {});
   }
 
   Future<String> _post(String path, [String body = '']) {
-    return _postImpl('${_baseUrl}/$path', const {}, body);
+    return _postImpl('$_baseUrl/$path', const {}, body);
   }
 
   Future<String> _delete(String path, [String body = '']) {
-    return _deleteImpl('${_baseUrl}/$path', const {}, body);
+    return _deleteImpl('$_baseUrl/$path', const {}, body);
   }
 
   Future<String> _put(String path, [String body = '']) {
-    return _putImpl('${_baseUrl}/$path', const {}, body);
+    return _putImpl('$_baseUrl/$path', const {}, body);
   }
 
   Future<String> _getImpl(String url, Map<String, String> headers) async {
-    return await _requestWrapper(
-        () async => await (await _client).get(url, headers: headers));
+    return await _requestWrapper(() async =>
+        await (await _client).get(Uri.parse(url), headers: headers));
   }
 
   Future<String> _postImpl(
       String url, Map<String, String> headers, dynamic body) async {
-    return await _requestWrapper(() async =>
-        await (await _client).post(url, headers: headers, body: body));
+    return await _requestWrapper(() async => await (await _client)
+        .post(Uri.parse(url), headers: headers, body: body));
   }
 
   Future<String> _deleteImpl(
@@ -146,8 +183,8 @@ abstract class SpotifyApiBase {
 
   Future<String> _putImpl(
       String url, Map<String, String> headers, dynamic body) async {
-    return await _requestWrapper(() async =>
-        await (await _client).put(url, headers: headers, body: body));
+    return await _requestWrapper(() async => await (await _client)
+        .put(Uri.parse(url), headers: headers, body: body));
   }
 
   Future<String> _requestWrapper(Future<http.Response> Function() request,
@@ -163,7 +200,7 @@ abstract class SpotifyApiBase {
         print(
             'Spotify API rate exceeded. waiting for ${ex.retryAfter} seconds');
         _shouldWait = true;
-        unawaited(Future.delayed(Duration(seconds: ex.retryAfter))
+        unawaited(Future.delayed(Duration(seconds: ex.retryAfter.toInt()))
             .then((v) => _shouldWait = false));
       }
     }
@@ -171,7 +208,7 @@ abstract class SpotifyApiBase {
   }
 
   Future<SpotifyApiCredentials> getCredentials() async {
-    return await SpotifyApiCredentials._fromClient(await _client);
+    return SpotifyApiCredentials._fromClient(await _client);
   }
 
   String handleErrors(http.Response response) {
@@ -181,7 +218,7 @@ abstract class SpotifyApiBase {
       final error = SpotifyError.fromJson(jsonMap['error']);
       if (response.statusCode == 429) {
         throw ApiRateException.fromSpotify(
-            error, num.parse(response.headers['retry-after']));
+            error, num.parse(response.headers['retry-after']!));
       }
       throw SpotifyException.fromSpotify(
         error,