+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog
 
+## 0.7.3
+
+- fix `isLast` in paging endpoint
+
+## 0.7.2
+
+- add update playlist endpoint
+- add docs for serverside auth token refresh
+
+## 0.7.1
+
+- several bug fixes
+- add async factory for `SpotifyApi`
+- add follow/unfollow artists and users endpoints
+- add inserting track in playlist endpoint
+- add remove tracks endpoint
+
+## 0.7.0
+
+- create client with access token only
+- add description field to playlist model
+
+## 0.6.0
+
+- fix following artist
+- add user saved shows endpoint
+- fix episode listing
+- clean up album models
+- add null safety
+
 ## 0.5.1
 
 - add createPlaylist optional parameters